replaced bulk_index.py, included progressbar; other minor fixes
